// apps/gb-ingest/src/gb-ingest.api-types.ts
export interface components {
    schemas: {
// ...
        IngestRequestInput: {
            source: {
                integration: string;
                device: string;
                instance?: string;
            };
            items: ({
                /** @constant */
                type: "sample";
                idempotency_key: string;
                metric: string;
                /** Format: date-time */
                start: string;
                /** Format: date-time */
                end: string;
                tz?: string;
                value: unknown;
                session_idempotency_key?: string;
            } | {
                /** @constant */
                type: "session";
                idempotency_key: string;
                session_type: string;
                /** Format: date-time */
                start: string;
                /** Format: date-time */
                end: string;
                tz?: string;
                metadata?: {
                    [key: string]: unknown;
                };
            } | {
                /** @constant */
                type: "event";
                idempotency_key: string;
                metric: string;
                /** Format: date-time */
                at: string;
                tz?: string;
                payload: {
                    [key: string]: unknown;
                };
                session_idempotency_key?: string;
            } | {
                /** @constant */
                type: "annotation";
                idempotency_key: string;
                /** Format: date-time */
                start: string;
                /** Format: date-time */
                end: string;
                tz?: string;
                text: string;
                tags?: string[];
            })[];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}